+++ docker/docker-compose.yml
@@ -1,0 +1,25 @@
+version: '3.8'
+
+services:
+  mysql:
+    image: mysql:8.0
+    container_name: online_bookstore
+    environment:
+      MYSQL_ROOT_PASSWORD: 1q2w3e4r
+      MYSQL_DATABASE: bookStore
+      MYSQL_USER: admin
+      MYSQL_PASSWORD: 1q2w3e4r
+    volumes:
+      - mysql_data:/var/lib/mysql  # MySQL 데이터 저장 볼륨
+      - ./my.cnf:/etc/mysql/conf.d/my.cnf  # MySQL UTF-8 설정 파일
+    ports:
+      - "3307:3306"
+    networks:
+      - online_network
+
+volumes:
+  mysql_data:
+
+networks:
+  online_network:
+    driver: bridge
