+++ src/main/java/com/bookstore/online/domain/review/repository/ReviewRepository.java
@@ -6,7 +6,6 @@ import com.bookstore.online.domain.review.entity.ReviewEntity;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import org.springframework.transaction.annotation.Transactional;
 
 @Repository
 public interface ReviewRepository extends JpaRepository<ReviewEntity, Integer> {
